+++ source/scripts/main.js
@@ -1,6 +1,7 @@
 // Uncomment if you need to support advanced ES6 features in IE11 and below
 // import 'babel-polyfill'
 
+import './utils/dom-polyfills'
 import $ from 'jquery'
 
 $('html').removeClass('no-js')